backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
aiohttp==3.9.3
